src/cloud_function/requirements.txt
# Function dependencies, for example:
# package>=version
flask
google-cloud-storage
numpy
joblib
scikit-learn